+++ CHANGELOG.md
@@ -30,6 +30,7 @@
 - (**BREAK**) changed `onTouchStart()`, `onTouchMove()` and `onTouchEnd()` callback signature to `(pos: Vec2, touch: Touch) => void` (exposes the native `Touch` object)
 - added `onResize()` to register an event that runs when canvas resizes
 - (**BREAK**) `GameObj#_id` has been renamed to `GameObj#id`
+- (**BREAK**) removed `load()` event for components, use `onLoad()` in `add()` event
 
 ### v2000.2.6